Split up common.h
