+++ src/kbmod/image_base/bindings.cpp
@@ -20,11 +20,6 @@ using std::to_string;
 PYBIND11_MODULE(search, m) {
     m.attr("KB_NO_DATA") = pybind11::float_(NO_DATA);
     m.attr("HAS_GPU") = pybind11::bool_(HAVE_GPU);
-    py::enum_<StampType>(m, "StampType")
-            .value("STAMP_SUM", StampType::STAMP_SUM)
-            .value("STAMP_MEAN", StampType::STAMP_MEAN)
-            .value("STAMP_MEDIAN", StampType::STAMP_MEDIAN)
-            .export_values();
     py::class_<pf>(m, "psf", py::buffer_protocol(), R"pbdoc(
             Point Spread Function.
 
@@ -176,34 +171,6 @@ PYBIND11_MODULE(search, m) {
             .def("get_width", &is::getWidth)
             .def("get_height", &is::getHeight)
             .def("get_npixels", &is::getNPixels);
-    py::class_<trajectory>(m, "trajectory", R"pbdoc(
-            A trajectory structure holding basic information about potential results.
-            )pbdoc")
-            .def(py::init<>())
-            .def_readwrite("x_v", &trajectory::xVel)
-            .def_readwrite("y_v", &trajectory::yVel)
-            .def_readwrite("lh", &trajectory::lh)
-            .def_readwrite("flux", &trajectory::flux)
-            .def_readwrite("x", &trajectory::x)
-            .def_readwrite("y", &trajectory::y)
-            .def_readwrite("obs_count", &trajectory::obsCount)
-            .def("__repr__",
-                 [](const trajectory &t) {
-                     return "lh: " + to_string(t.lh) + " flux: " + to_string(t.flux) +
-                            " x: " + to_string(t.x) + " y: " + to_string(t.y) + " x_v: " + to_string(t.xVel) +
-                            " y_v: " + to_string(t.yVel) + " obs_count: " + to_string(t.obsCount);
-                 })
-            .def(py::pickle(
-                    [](const trajectory &p) {  // __getstate__
-                        return py::make_tuple(p.xVel, p.yVel, p.lh, p.flux, p.x, p.y, p.obsCount);
-                    },
-                    [](py::tuple t) {  // __setstate__
-                        if (t.size() != 7) throw std::runtime_error("Invalid state!");
-                        trajectory trj = {t[0].cast<float>(), t[1].cast<float>(), t[2].cast<float>(),
-                                          t[3].cast<float>(), t[4].cast<short>(), t[5].cast<short>(),
-                                          t[6].cast<short>()};
-                        return trj;
-                    }));
     py::class_<pixelPos>(m, "pixel_pos")
             .def(py::init<>())
             .def_readwrite("x", &pixelPos::x)
@@ -217,17 +184,4 @@ PYBIND11_MODULE(search, m) {
             .def_readwrite("m11", &imageMoments::m11)
             .def_readwrite("m02", &imageMoments::m02)
             .def_readwrite("m20", &imageMoments::m20);
-    py::class_<stampParameters>(m, "stamp_parameters")
-            .def(py::init<>())
-            .def_readwrite("radius", &stampParameters::radius)
-            .def_readwrite("stamp_type", &stampParameters::stamp_type)
-            .def_readwrite("do_filtering", &stampParameters::do_filtering)
-            .def_readwrite("center_thresh", &stampParameters::center_thresh)
-            .def_readwrite("peak_offset_x", &stampParameters::peak_offset_x)
-            .def_readwrite("peak_offset_y", &stampParameters::peak_offset_y)
-            .def_readwrite("m01", &stampParameters::m01_limit)
-            .def_readwrite("m10", &stampParameters::m10_limit)
-            .def_readwrite("m11", &stampParameters::m11_limit)
-            .def_readwrite("m02", &stampParameters::m02_limit)
-            .def_readwrite("m20", &stampParameters::m20_limit);
 }
